# infrastructure/api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      csrf_token:
        type: string
      refresh_token:
        type: string
      user: {}
    type: object
...
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  handlers.RegisterResponse:
    properties:
      access_token:
        type: string
      csrf_token:
        type: string
      refresh_token:
        type: string
      user: {}
    type: object
...
host: api.felix-freund.com
info:
  contact:
    email: support@api.felix-freund.com
    name: API Support
    url: https://api.felix-freund.com/support
  description: Secure file storage and management API with authentication, email verification,
    and password reset.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://api.felix-freund.com/terms
  title: NAS.AI API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with tokens
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Authentication
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email verification
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully with tokens
          schema:
            $ref: '#/definitions/handlers.RegisterResponse'
        "400":
          description: Invalid request or weak password
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email already registered
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  CSRFToken:
    description: CSRF token for state-changing operations
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"